+++ iosHyperskillApp/iosHyperskillApp/Sources/Modules/App/AppViewModel.swift
@@ -92,6 +92,7 @@ final class AppViewModel: FeatureViewModel<AppFeatureState, AppFeatureMessage, A
 
 extension AppViewModel: AuthOutputProtocol {
     func handleUserAuthorized(profile: Profile) {
+        #warning("ALTAPPS-971: Provide isNotificationPermissionGranted")
         onNewMessage(
             AppFeatureMessageUserAuthorized(
                 profile: profile,
