+++ iosHyperskillApp/iosHyperskillApp/Sources/Modules/App/ViewControllers/AppViewController.swift
@@ -113,7 +113,7 @@ extension AppViewController: AppViewControllerProtocol {
                 navigationController.navigationBar.prefersLargeTitles = true
                 return navigationController
             case .notificationOnBoardingScreen:
-                // TODO: ALTAPPS-971
+                #warning("ALTAPPS-971: Route to the NotificationOnBoardingScreen")
                 fatalError("NotificationOnBoardingScreen is not implemented yet!")
             }
         }()
