+++ shared/src/androidMain/kotlin/org/hyperskill/app/core/injection/CommonAndroidAppGraphImpl.kt
@@ -16,8 +16,8 @@ import org.hyperskill.app.home.injection.PlatformHomeComponent
 import org.hyperskill.app.home.injection.PlatformHomeComponentImpl
 import org.hyperskill.app.notification.remote.injection.AndroidPlatformPushNotificationsPlatformDataComponent
 import org.hyperskill.app.notification.remote.injection.PlatformPushNotificationsDataComponent
-import org.hyperskill.app.notification_onboarding.injection.PlatformNotificationsOnboardingComponent
-import org.hyperskill.app.notification_onboarding.injection.PlatformNotificationsOnboardingComponentImpl
+import org.hyperskill.app.notifications_onboarding.injection.PlatformNotificationsOnboardingComponent
+import org.hyperskill.app.notifications_onboarding.injection.PlatformNotificationsOnboardingComponentImpl
 import org.hyperskill.app.onboarding.injection.OnboardingComponent
 import org.hyperskill.app.onboarding.injection.PlatformOnboardingComponent
 import org.hyperskill.app.onboarding.injection.PlatformOnboardingComponentImpl
