--- shared/src/androidMain/kotlin/org/hyperskill/app/notification_onboarding/presentation/NotificationOnboardingViewModel.kt
@@ -1,11 +1,0 @@
-package org.hyperskill.app.notification_onboarding.presentation
-
-import org.hyperskill.app.notification_onboarding.presentation.NotificationsOnboardingFeature.Action.ViewAction
-import org.hyperskill.app.notification_onboarding.presentation.NotificationsOnboardingFeature.Message
-import org.hyperskill.app.notification_onboarding.presentation.NotificationsOnboardingFeature.State
-import ru.nobird.android.view.redux.viewmodel.ReduxViewModel
-import ru.nobird.app.presentation.redux.container.ReduxViewContainer
-
-class NotificationOnboardingViewModel(
-    reduxViewContainer: ReduxViewContainer<State, Message, ViewAction>
-) : ReduxViewModel<State, Message, ViewAction>(reduxViewContainer)
