+++ shared/src/commonMain/kotlin/org/hyperskill/app/main/presentation/AppFeature.kt
@@ -28,7 +28,7 @@ interface AppFeature {
         @Serializable
         data class Ready(
             val isAuthorized: Boolean,
-            val profile: Profile? = null
+            internal val profile: Profile? = null
         ) : State
     }
