+++ shared/src/commonMain/kotlin/org/hyperskill/app/notifications_onboarding/injection/NotificationsOnboardingFeatureBuilder.kt
@@ -1,18 +1,18 @@
-package org.hyperskill.app.notification_onboarding.injection
+package org.hyperskill.app.notifications_onboarding.injection
 
 import org.hyperskill.app.analytic.domain.interactor.AnalyticInteractor
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
 import org.hyperskill.app.notification.local.domain.interactor.NotificationInteractor
-import org.hyperskill.app.notification_onboarding.presentation.NotificationsOnboardingActionDispatcher
-import org.hyperskill.app.notification_onboarding.presentation.NotificationsOnboardingFeature.Action
-import org.hyperskill.app.notification_onboarding.presentation.NotificationsOnboardingFeature.Message
-import org.hyperskill.app.notification_onboarding.presentation.NotificationsOnboardingFeature.State
-import org.hyperskill.app.notification_onboarding.presentation.NotificationsOnboardingReducer
+import org.hyperskill.app.notifications_onboarding.presentation.NotificationsOnboardingActionDispatcher
+import org.hyperskill.app.notifications_onboarding.presentation.NotificationsOnboardingFeature.Action
+import org.hyperskill.app.notifications_onboarding.presentation.NotificationsOnboardingFeature.Message
+import org.hyperskill.app.notifications_onboarding.presentation.NotificationsOnboardingFeature.State
+import org.hyperskill.app.notifications_onboarding.presentation.NotificationsOnboardingReducer
 import ru.nobird.app.presentation.redux.dispatcher.wrapWithActionDispatcher
 import ru.nobird.app.presentation.redux.feature.Feature
 import ru.nobird.app.presentation.redux.feature.ReduxFeature
 
-object NotificationsOnboardingFeatureBuilder {
+internal object NotificationsOnboardingFeatureBuilder {
     fun build(
         analyticInteractor: AnalyticInteractor,
         notificationInteractor: NotificationInteractor
