+++ shared/src/commonMain/kotlin/org/hyperskill/app/notifications_onboarding/presentation/NotificationsOnboardingFeature.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.notification_onboarding.presentation
+package org.hyperskill.app.notifications_onboarding.presentation
 
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
 
@@ -13,12 +13,14 @@ object NotificationsOnboardingFeature {
     }
 
     sealed interface Action {
-
-        data class LogAnalyticsEvent(val event: HyperskillAnalyticEvent) : Action
-        object UpdateLastNotificationPermissionRequestTime : Action
         sealed interface ViewAction : Action {
             object RequestNotificationPermission : ViewAction
             object CompleteNotificationOnboarding : ViewAction
         }
+    }
+
+    internal sealed interface InternalAction : Action {
+        data class LogAnalyticsEvent(val event: HyperskillAnalyticEvent) : Action
+        object UpdateLastNotificationPermissionRequestTime : Action
     }
 }
